refactor(Cozinha) ISS-013: refactor to organization
