+++ src/main/java/com/algaworks/algafood/jpa/cozinha/AlteracaoCozinhaMain.java
@@ -1,4 +1,4 @@
-package com.algaworks.algafood.jpa;
+package com.algaworks.algafood.jpa.cozinha;
 
 import com.algaworks.algafood.AlgafoodApiApplication;
 import com.algaworks.algafood.domain.model.Cozinha;
@@ -7,7 +7,7 @@ import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
 
-public class BuscaCozinhaMain {
+public class AlteracaoCozinhaMain {
 
     public static void main(String[] args) {
         // Configuração que faz a api rodar e depois já finalizar, para podermos realizar alguns testes
@@ -18,11 +18,11 @@ public class BuscaCozinhaMain {
         // Pegamos o Bean que inicializamos no Cadastro cozinha para podermos realizar as consultas
         CozinhaRepository cozinhaRepository = applicationContext.getBean(CozinhaRepository.class);
 
-        // Buscando cozinha com id 1
-        Cozinha cozinha = cozinhaRepository.buscarPorId(1L);
+        Cozinha cozinha = new Cozinha();
+        cozinha.setId(1L);
+        cozinha.setNome("Brasileira");
 
-        System.out.println(cozinha.getNome());
-
+        cozinhaRepository.salvar(cozinha);
     }
 
 }
